third_party/WebKit/LayoutTests/inspector/sources/debugger/debugger-return-value.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/debugger-test.js"></script>
<script>

function d()
{
    var i = 10;
    return i;
}

function testFunction()
{
    debugger;
    return d();
}

var test = function()
{
    InspectorTest.setQuiet(true);
    InspectorTest.startDebuggerTest(step1);

    function step1()
    {
        InspectorTest.runTestFunctionAndWaitUntilPaused(step2);
    }

    var stepCount = 0;
    function step2(callFrames)
    {
        if (stepCount < 2) {
            for (var i = 0, frame; frame = callFrames[i]; ++i)
                InspectorTest.assertTrue(!frame.returnValue(), "Unexpected returnValue in frame #" + i);
            WebInspector.panels.sources._stepOverButton.element.click();
            InspectorTest.waitUntilResumed(InspectorTest.waitUntilPaused.bind(InspectorTest, step2));
        } else {
            InspectorTest.captureStackTrace(callFrames, null, { printReturnValue: true });
            InspectorTest.completeDebuggerTest();
        }
        ++stepCount;
    }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests function's return value reported from backend.
</p>

</body>
</html>
